requirements.txt
streamlit==1.29.0
numpy==1.24.4
pandas==2.0.3
opencv-python-headless==4.8.0.76
pillow==10.0.1
google-cloud-storage==2.13.0
torch==2.0.1
torchvision==0.15.2
ultralytics==8.0.196
watchdog==3.0.0
